Implement DetectModelFromLogsView for log analysis and model detection
- Added DetectModelFromLogsView to scan container logs for vLLM startup patterns and return detected model information, including hf_model_id, model_type, and port.
- Introduced helper functions for parsing logs and inferring model types based on the detected hf_model_id.
- Updated URL patterns to include a new endpoint for detecting models from logs, enhancing the backend's capabilities for model management.
