+++ app/backend/docker_control/urls.py
@@ -28,6 +28,7 @@ from .views import (
     DiscoverContainersView,
     RegisterExternalModelView,
     AvailableDevicesView,
+    DetectModelFromLogsView,
 )
 
 urlpatterns = [
@@ -53,4 +54,5 @@ urlpatterns = [
     path("discover-containers/", views.DiscoverContainersView.as_view(), name="discover-containers"),
     path("register-external/", views.RegisterExternalModelView.as_view(), name="register-external"),
     path("available-devices/", views.AvailableDevicesView.as_view(), name="available-devices"),
+    path("detect-model/<str:container_id>/", views.DetectModelFromLogsView.as_view(), name="detect-model"),
 ]
