+++ app/frontend/package.json
@@ -58,7 +58,7 @@
     "cmdk": "^1.0.4",
     "compromise": "^14.14.4",
     "framer-motion": "^11.18.2",
-    "lucide-react": "^0.344.0",
+    "lucide-react": "^0.468.0",
     "mini-svg-data-uri": "^1.4.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
