+++ app/frontend/src/components/NavBar.tsx
@@ -21,6 +21,7 @@ import {
   Cog,
   Menu,
   History,
+  PackagePlus,
 } from "lucide-react";
 
 import { useLogo } from "../utils/logo";
@@ -498,6 +499,13 @@ export default function NavBar() {
       icon: Boxes,
       label: "Models Deployed",
       tooltip: "Manage deployed models",
+    },
+    {
+      type: "link",
+      to: "/register-model",
+      icon: PackagePlus,
+      label: "Register Model",
+      tooltip: "Register an externally running model container",
     },
     {
       type: "link",
