+++ app/frontend/src/routes/route-config.tsx
@@ -56,6 +56,7 @@ import SpeechToTextPage from "../pages/SpeechToTextPage";
 import ApiInfoPage from "../pages/ApiInfoPage";
 import DeploymentHistoryPage from "../pages/DeploymentHistoryPage";
 import TTSPage from "../pages/TTSPage";
+import RegisterModelPage from "../pages/RegisterModelPage";
 
 // Define route configuration type
 export interface RouteConfig {
@@ -75,6 +76,11 @@ export const getRoutes = (): RouteConfig[] => {
     {
       path: "/models-deployed",
       element: <ModelsDeployed />,
+      condition: true,
+    },
+    {
+      path: "/register-model",
+      element: <RegisterModelPage />,
       condition: true,
     },
     {
